django_backend/requirements.txt
Django>=4.2,<6.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.0
firebase-admin>=6.2.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
sentry-sdk>=1.40.0
celery>=5.3.6
redis>=5.0.1
stripe>=7.0.0
requests>=2.31.0
reportlab>=4.0.0
openpyxl>=3.1.2
